requirements.txt
azure-functions==1.11.2
postal==1.1.10